Prevent implicit type promotion to double precision

+++ Code/Teensy/src/Filter/Feedback.cpp
@@ -15,17 +15,17 @@ float getFeedbackErrorTerm(float current) {
   // Cleanly average -5 pA noise band when setpoint is 10 pA
   float clampedCurrent;
   if (Feedback::setpointVoltage >= 0) {
-    clampedCurrent = max(current, -5e-12);
+    clampedCurrent = max(current, -5e-12f);
   } else {
-    clampedCurrent = min(current, 5e-12);
+    clampedCurrent = min(current, 5e-12f);
   }
 
   float ratio = clampedCurrent / expectedCurrent;
-  float k = 1.025e10 * sqrt(Feedback::tunnelingBarrierHeight);
-  float kΔz = ratio - 1;
+  float k = 1.025e10f * sqrtf(Feedback::tunnelingBarrierHeight);
+  float kΔz = ratio - 1.0f;
 
   // Make it even more repulsive at high currents.
-  float F = exp(k * 50e-12);
+  float F = expf(k * 50e-12f);
   if (ratio > F) {
     kΔz += (ratio - F) * (ratio - F);
   }
@@ -51,10 +51,10 @@ float Feedback::getVoltageCorrection(float current) {
   // enforcing a limit on approach speed will not change anything
   float timeProgress = float(KilohertzLoop::period) / float(integratorTimeLag);
   float correctionInMeters = -dz * timeProgress;
-  float correctionInVolts = correctionInMeters / 0.320e-9;
+  float correctionInVolts = correctionInMeters / 0.320e-9f;
 
   // Limit slew rate to 0.5 V/μs.
-  float maxVoltageChange = 0.5 * float(KilohertzLoop::period);
+  float maxVoltageChange = 0.5f * float(KilohertzLoop::period);
   correctionInVolts = max(correctionInVolts, -maxVoltageChange);
   correctionInVolts = min(correctionInVolts, maxVoltageChange);
   return correctionInVolts;
